+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/AutoResizableTextareaMobileOnTouchStart.test.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Textarea from '../../../Textarea';
 
 export default class textarea extends React.Component {
-  componentWillMount() {
+  componentDidMount() {
     if (!window.ontouchstart) {
       this.resetontouchstart = true;
       window.ontouchstart = 'true';