+++ packages/terra-time-input/src/terra-dev-site/test/time-input/twelve-hour/Mobile.test.jsx
@@ -8,7 +8,7 @@ export default class TimeInputElement extends React.Component {
     this.handleOnChange = this.handleOnChange.bind(this);
   }
 
-  componentWillMount() {
+  componentDidMount() {
     if (!window.ontouchstart) {
       this.resetontouchstart = true;
       window.ontouchstart = 'true';