+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed use of componentWillMount from examples
 
 2.33.0 - (November 13, 2018)
 ------------------